refactor the search bar to be more generic to support search on other data
